--- module/Application/view/application/index/index.phtml
@@ -1,78 +1,0 @@
-<!--<div class="row">
-
-    <div class="col-md-4">
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                <h3 class="panel-title">Portal Test Info</h3>
-            </div>
-            <div class="panel-body">
-                <p>
-                   <?php if ($this->newSession)
-                   {
-                   	echo "New Session<br>";
-                   }
-                   else
-                   {
-                   	echo "Existing Session<br>";
-                   }
-                   echo "User ID: " . $this->userId;
-                   echo "<br>Session Level: " . $this->sessionLevel;
-                   echo "<br>Active Time: " . date("Y-m-d H:i:s", $this->activeTime); ?>
-                </p>
-
-
-            </div>
-        </div>
-    </div>
-</div>-->
-
-<div class="row">
-    <div class="col-xs-6 col-sm-3 col-md-2">
-        <div class="app-tile">
-            <a href="https://drive.google.com" target="_blank">
-                <img src="http://icons.iconarchive.com/icons/marcus-roberto/google-play/512/Google-Drive-icon.png" alt="">
-                <span class="app-name">Google Drive</span>
-            </a>
-        </div>
-    </div>
-    <div class="col-xs-6 col-sm-3 col-md-2">
-        <div class="app-tile">
-            <a href="https://docs.google.com" target="_blank">
-                <img src="https://lh4.ggpht.com/-wROmWQVYTcjs3G6H0lYkBK2nPGYsY75Ik2IXTmOO2Oo0SMgbDtnF0eqz-BRR1hRQg=w300" alt="">
-                <span class="app-name">Google Docs</span>
-            </a>
-        </div>
-    </div>
-    <div class="col-xs-6 col-sm-3 col-md-2">
-        <div class="app-tile">
-            <a href="https://digitalocean.com" target="_blank">
-                <img src="https://logos.bugcrowdusercontent.com/logos/b71d/c1f5/1f8f12c4/a5e0c009d691e9be46fbcaed77c9cc4f_DO_Logo_icon_blue-larger.png" alt="">
-                <span class="app-name">Digital Ocean</span>
-            </a>
-        </div>
-    </div>
-    <div class="col-xs-6 col-sm-3 col-md-2">
-        <div class="app-tile">
-            <a href="https://gmail.com" target="_blank">
-                <img src="https://upload.wikimedia.org/wikipedia/commons/4/4e/Gmail_Icon.png" alt="">
-                <span class="app-name">Gmail</span>
-            </a>
-        </div>
-    </div>
-    <div class="col-xs-6 col-sm-3 col-md-2">
-        <div class="app-tile">
-            <a href="https://unomaha.edu" target="_blank">
-                <img src="https://www.unomaha.edu/news/img/0161-IMG_MDMW_UC_TERMINOLOGY2.jpg" alt="">
-                <span class="app-name">UNOmaha</span>
-            </a>
-        </div>
-    </div>
-    <div class="col-xs-6 col-sm-3 col-md-2">
-        <div class="app-tile">
-            <a href="https://ubuntu.com" target="_blank">
-                <img src="http://icons.iconarchive.com/icons/dakirby309/simply-styled/256/OS-Ubuntu-icon.png" alt="">
-                <span class="app-name">Ubuntu</span>
-            </a>
-        </div>
-    </div>
-</div>
